Initial work.
Harness heavily based on the 1kusai one.

Drawing is up and running, but I need to simplify some of the shapes
still. Then on with adding the music! I should have  about 200 bytes
(after packing) left, let's see how far we can go with that.

It may be possible to free more by using "set origin" instead of (or in
addition to "move absolute" to start the shapes. Then the 4 last ones would
be even more identical to each other, and all coordinates could fit on 8 bits
which would save a lot of space.